pyproject.toml
[project]
name = "bu-banking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "asgiref==3.11.1",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.5",
    "coreapi==2.3.3",
    "coreschema==0.0.4",
    "django==6.0.3",
    "django-cors-headers==4.9.0",
    "django-rest-authtoken==2.1.4",
    "django-rest-swagger==2.2.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "dotenv>=0.9.9",
    "drf-yasg==1.21.15",
    "idna==3.11",
    "inflection==0.5.1",
    "itypes==1.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.3",
    "openapi-codec==1.3.2",
    "packaging==26.0",
    "pycparser==3.0",
    "pyjwt==2.11.0",
    "pytz==2026.1.post1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "simplejson==3.20.2",
    "sqlparse==0.5.5",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
]
